temp save to remove server.go and jwt.go
