+++ core/base.go
@@ -1,8 +1,11 @@
 package core
 
 import (
+	"time"
+
 	"github.com/Z-M-Huang/Tools/data"
 	"github.com/Z-M-Huang/Tools/utils"
+	"github.com/gin-gonic/gin"
 )
 
 //GetResponseInContext get response struct from context
@@ -14,11 +17,11 @@ func GetResponseInContext(contextKey map[string]interface{}) *data.Response {
 	return response.(*data.Response)
 }
 
-//GetClaimInContext get claim struct from context
-func GetClaimInContext(contextKey map[string]interface{}) *data.JWTClaim {
-	claim := contextKey[utils.ClaimCtxKey]
-	if claim == nil {
-		return nil
+//SetCookie sets cookie
+func SetCookie(c *gin.Context, cookieName, cookieVal string, expiresAt time.Time, httpOnly bool) {
+	if data.Config.HTTPS {
+		c.SetCookie(cookieName, cookieVal, int(expiresAt.Sub(time.Now()).Seconds()), "/", data.Config.Host, true, httpOnly)
+	} else {
+		c.SetCookie(cookieName, cookieVal, int(expiresAt.Sub(time.Now()).Seconds()), "/", data.Config.Host, false, httpOnly)
 	}
-	return claim.(*data.JWTClaim)
 }
